requirements.txt
numpy ~= 1.25.2
pygame ~= 2.5.1
pillow ~= 10.0.0
